run.sh: relax regexp to have it work in Docker (fixes #1)
The version of `yarn` in this popular Hadoop Dockere image does not
produce whitespace at the beginning of the lines with RUNNING. The
original regexp used to compute `MAPPERS` requires this whitespace,
incorrectly asking for 0 mappers and producing an
IntegerFormatException.

This commit relaxes \s+ to \s* in the regexp so it works again.